client-api: Move ProfileFieldName and ProfileFieldValue to ruma-common
To be able to reuse them in the federation API.
